+++ Sources/SF2Lib/include/SF2Lib/DSP/DSP.hpp
@@ -105,14 +105,14 @@ inline constexpr Float tenthPercentageToNormalized(Float value) noexcept {
  */
 extern void panLookup(Float pan, Float& left, Float& right) noexcept;
 
-extern double centsPartialLookup(int partial) noexcept;
+extern Float centsPartialLookup(int partial) noexcept;
 
 /**
  Quickly convert cent value into a frequency using a table lookup. These calculations are taken from the Fluid Synth
  fluid_conv.c file, in particular the fluid_ct2hz_real function. Uses CentPartialLookup above to convert values from
  0 - 1199 into the proper multiplier.
  */
-inline double centsToFrequency(Float value) noexcept {
+inline Float centsToFrequency(Float value) noexcept {
   if (value < 0.0f) [[unlikely]] return 1.0f;
 
   // This seems to be the fastest way to do the following. Curiously, the operation `cents % 1200` is faster than doing
