+++ Sources/SF2Lib/include/SF2Lib/Render/Voice/Sample/Index.hpp
@@ -12,7 +12,7 @@ namespace SF2::Render::Voice::Sample {
 
 /**
  Interpolatable index into a NormalizedSampleSource. Maintains two counters, an integral one (`size_t`) and a partial
- one (`double`) that indicates how close the index is to a sample index. These two values are then used by other
+ one (`Float`) that indicates how close the index is to a sample index. These two values are then used by other
  routines to fetch the appropriate samples and interpolate over them.
 
  Updates to the index honor loops in the sample stream if allowed. The index can also signal when it has reached the
